Make rotation and scale to work again

+++ Digital Album/Application/Albums/AlbumViewController.m
@@ -232,7 +232,7 @@
     DAPage * page = [[DAPage alloc] init];
     self.album.pages = [self.album.pages arrayByAddingObject:page];
     
-    NSArray * array = @[[self pageControllerAtIndex:self.album.pages.count - 1]];
+    NSArray * array = @[[self pageControllerAtIndex:(int)self.album.pages.count - 1]];
     [self.pageViewController setViewControllers:array direction:UIPageViewControllerNavigationDirectionForward animated:YES completion:nil];
     
     [self donePressed];
